style(telemetry): satisfy phpcs on metrics code
Extract FakeMetrics test double to tests/Support/, add missing docblocks,
import Sentry\traceMetrics via use statement, and normalize declare
statement spacing.
