+++ backend/tests/Feature/TelemetryMetricsBindingTest.php
@@ -1,5 +1,4 @@
 <?php
-
 declare(strict_types=1);
 
 namespace Tests\Feature;
